requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-yasg>=1.21
requests>=2.31
